removed duplicates and cleaned up the code
